phase5_langgraph_bifrost/src/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langgraph==0.0.26
langchain-core==0.1.28
langsmith==0.1.13
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-langchain==0.0.1
requests==2.31.0
langchain-openai==0.0.8
